♻️ Move protocol delete modal to a separate file
